Model: add color() member function
This returns the difference in magnitudes between two filters. Even
though it simply invokes Model.bandmag() twice, we require that the two
bandpass arguments be scalars, since lists of pairs (or pairs of lists)
of filters or magnitude systems leads to cumbersome output. The only
allowable iterable is the "phase" argument.

Signed-off-by: Brian Friesen <friesen@ou.edu>