update CI to use external extinction module
In the process, I removed a configuration from appveyor. With Python 3.4
and numpy 1.9, pip-installing `extinction` resulted in an error that seemed
to have to do with looking for MSVC 9 but having MSVC 10 installed via
conda. This was very confusing because the `extinction` tests use the same
configuration and pass. Could be a difference in conda packages
due to installing cython or not, but it seemed like a bug elsewhere and
not worth hunting down for us.